worked on refactoring architecture app
