+++ src/api/posts/get/getPosts.js
@@ -1,25 +1,17 @@
 //imports
 import axios, * as others from 'axios'
-import { ref } from 'vue'
 
-//data
-export const loadingData = ref(false)
-export const preLoaderStatus = ref(false)
 
 //methods
 export async function getAllPosts () {
   const SERVER_URL = 'http://localhost:3000/posts'
-  preLoaderStatus.value = true
   return axios
     .get(SERVER_URL)
     .then(Response => {
-      preLoaderStatus.value = false
       return Response.data
     })
     .catch(Error => {
       console.log(Error)
-      preLoaderStatus.value = false
-      loadingData.value = true
-      return Error.response.data.Error
+      return Error.response
     })
 }
